Add rendering module
